app/requirements.txt
django>=2.0,<2.1
django-filter>=1.1,<1.2
djangorestframework>=3.8,<3.9
pillow==7.0.0
whitenoise==5.0.1
django-heroku==0.3.1
gunicorn==20.0.4
dj-database-url==0.5.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt
